+++ src/platform.ts
@@ -47,6 +47,7 @@ export class HoneywellHomePlatform implements DynamicPlatformPlugin {
 
   public sensorData = [];
   private refreshInterval;
+  debugMode!: boolean;
 
   constructor(public readonly log: Logger, public readonly config: HoneywellPlatformConfig, public readonly api: API) {
     this.log.debug('Finished initializing platform:', this.config.name);
@@ -65,11 +66,13 @@ export class HoneywellHomePlatform implements DynamicPlatformPlugin {
       return;
     }
 
+    this.debugMode = process.argv.includes('-D') || process.argv.includes('--debug');
+    
     // setup axios interceptor to add headers / api key to each request
     this.axios.interceptors.request.use((request: AxiosRequestConfig) => {
-      request.headers.Authorization = `Bearer ${this.config.credentials ?.accessToken}`;
+      request.headers.Authorization = `Bearer ${this.config.credentials?.accessToken}`;
       request.params = request.params || {};
-      request.params.apikey = this.config.credentials ?.consumerKey;
+      request.params.apikey = this.config.credentials?.consumerKey;
       request.headers['Content-Type'] = 'application/json';
       return request;
     });
@@ -120,43 +123,43 @@ export class HoneywellHomePlatform implements DynamicPlatformPlugin {
 
     this.config.options = this.config.options || {};
 
-    if (this.config.options ?.thermostat) {
-      // Thermostat Config Options
-      this.config.options.thermostat.hide;
-      this.config.options.thermostat.hide_fan;
-      this.config.options.thermostat.thermostatSetpointStatus =
-        this.config.options.thermostat.thermostatSetpointStatus || 'PermanentHold';
-    }
-
-    if (this.config.options ?.leaksensor) {
-      // Leak Sensor Config Options
-      this.config.options.leaksensor.hide;
-      this.config.options.leaksensor.hide_humidity;
-      this.config.options.leaksensor.hide_temperature;
-      this.config.options.leaksensor.hide_leak;
-    }
-
-    if (this.config.options ?.roomsensor) {
-      // Room Sensor Config Options
-      this.config.options.roomsensor.hide;
-      this.config.options.roomsensor.hide_temperature;
-      this.config.options.roomsensor.hide_occupancy;
-      this.config.options.roomsensor.hide_humidity;
-    }
-
-    if (this.config.options ?.roompriority) {
-      // Room Priority Config Options
-      this.config.options.roompriority.thermostat;
-      this.config.options.roompriority.priorityType = this.config.options.roompriority.priorityType || 'PickARoom';
-    }
-
     if (this.config.options) {
-      this.config.options.ttl = this.config.options!.ttl || 300; // default 300 seconds
-    }
+      // Hide Devices by DeviceID
+      this.config.options.hide_device = this.config.options.hide_device || [];
 
-    if (!this.config.credentials ?.consumerSecret && this.config.options!.ttl! < 300) {
-      this.log.debug('TTL must be set to 300 or higher unless you setup your own consumerSecret.');
-      this.config.options!.ttl! = 300;
+      if (this.config.options?.thermostat) {
+        // Thermostat Config Options
+        this.config.options.thermostat.hide_fan;
+        this.config.options.thermostat.thermostatSetpointStatus =
+          this.config.options.thermostat.thermostatSetpointStatus || 'PermanentHold';
+      }
+
+      if (this.config.options?.leaksensor) {
+        // Leak Sensor Config Options
+        this.config.options.leaksensor.hide_humidity;
+        this.config.options.leaksensor.hide_temperature;
+        this.config.options.leaksensor.hide_leak;
+      }
+
+      if (this.config.options?.roomsensor) {
+        // Room Sensor Config Options
+        this.config.options.roomsensor.hide_temperature;
+        this.config.options.roomsensor.hide_occupancy;
+        this.config.options.roomsensor.hide_humidity;
+      }
+
+      if (this.config.options?.roompriority) {
+        // Room Priority Config Options
+        this.config.options.roompriority.thermostat;
+        this.config.options.roompriority.priorityType = this.config.options.roompriority.priorityType || 'PickARoom';
+      }
+
+      this.config.options.refreshRate = this.config.options!.refreshRate || 300; // default 300 seconds
+
+      if (!this.config.credentials?.consumerSecret && this.config.options!.refreshRate! < 300) {
+        this.log.debug('Refresh Rate must be set to 300 or higher unless you setup your own consumerSecret.');
+        this.config.options!.refreshRate! = 300;
+      }
     }
 
     if (!this.config.credentials) {
@@ -221,12 +224,16 @@ export class HoneywellHomePlatform implements DynamicPlatformPlugin {
         }
       }
 
+      
       this.config.credentials!.accessToken = result.access_token;
-      this.log.warn('Got access token:', this.config.credentials!.accessToken);
-
+      if (this.debugMode) {
+        this.log.warn('Got access token:', this.config.credentials!.accessToken);
+      }
       // check if the refresh token has changed
       if (result.refresh_token !== this.config.credentials!.refreshToken) {
-        this.log.warn('New refresh token:', result.refresh_token);
+        if (this.debugMode) {
+          this.log.warn('New refresh token:', result.refresh_token);
+        }
         await this.updateRefreshToken(result.refresh_token);
       }
 
@@ -274,8 +281,9 @@ export class HoneywellHomePlatform implements DynamicPlatformPlugin {
 
       // save the config, ensuring we maintain pretty json
       writeFileSync(this.api.user.configPath(), JSON.stringify(currentConfig, null, 4));
-
-      this.log.warn('Homebridge config.json has been updated with new refresh token.');
+      if (this.debugMode) {
+        this.log.warn('Homebridge config.json has been updated with new refresh token.');
+      }
     } catch (e) {
       this.log.error('Failed to update refresh token in config:', JSON.stringify(e.message));
       this.log.debug(JSON.stringify(e));
@@ -306,7 +314,14 @@ export class HoneywellHomePlatform implements DynamicPlatformPlugin {
         data: this.normalizeSensorDate(response.data),
       };
     } else {
-      this.log.info('getCurrentSensorData Cache %s %s - %s', device.deviceType, device.deviceModel, device.userDefinedDeviceName);
+      if (this.debugMode) {
+        this.log.info(
+          'getCurrentSensorData Cache %s %s - %s',
+          device.deviceType,
+          device.deviceModel,
+          device.userDefinedDeviceName,
+        );
+      }
     }
     return this.sensorData[device.deviceID].data;
   }
@@ -330,17 +345,26 @@ export class HoneywellHomePlatform implements DynamicPlatformPlugin {
     if (device.deviceID.startsWith('LCC') && device.deviceModel.startsWith('T9') && device.groups) {
       for (const group of device.groups) {
         const roomsensors = await this.getCurrentSensorData(device, group, locationId);
-        if (this.config.options ?.roompriority ?.thermostat) {
+        if (this.config.options?.roompriority?.thermostat) {
           this.log.info(`Total Rooms Found: ${roomsensors.length}`);
         }
         for (const accessories of roomsensors) {
           if (accessories) {
             for (const key in accessories) {
               const sensorAccessory = accessories[key];
-              if (sensorAccessory.accessoryAttribute && sensorAccessory.accessoryAttribute.type 
-                && sensorAccessory.accessoryAttribute.type.startsWith('Thermostat')) {
-                this.log.debug('Software Revision', group.id, sensorAccessory.roomId, sensorAccessory.accessoryId, 
-                  sensorAccessory.accessoryAttribute.name, JSON.stringify(sensorAccessory.accessoryAttribute.softwareRevision));
+              if (
+                sensorAccessory.accessoryAttribute &&
+                sensorAccessory.accessoryAttribute.type &&
+                sensorAccessory.accessoryAttribute.type.startsWith('Thermostat')
+              ) {
+                this.log.debug(
+                  'Software Revision',
+                  group.id,
+                  sensorAccessory.roomId,
+                  sensorAccessory.accessoryId,
+                  sensorAccessory.accessoryAttribute.name,
+                  JSON.stringify(sensorAccessory.accessoryAttribute.softwareRevision),
+                );
                 return sensorAccessory.accessoryAttribute.softwareRevision;
               } else {
                 this.log.info('No Thermostat', device, group, locationId);
@@ -352,7 +376,12 @@ export class HoneywellHomePlatform implements DynamicPlatformPlugin {
         }
       }
     } else {
-      this.log.info('Not a T9 LCC', device.deviceID.startsWith('LCC'), device.deviceModel.startsWith('T9'), device.groups);
+      this.log.info(
+        'Not a T9 LCC',
+        device.deviceID.startsWith('LCC'),
+        device.deviceModel.startsWith('T9'),
+        device.groups,
+      );
     }
   }
 
@@ -381,31 +410,63 @@ export class HoneywellHomePlatform implements DynamicPlatformPlugin {
                   this.log.error('Failed to Get T9 Firmware Version.', JSON.stringify(e.message));
                   this.log.debug(JSON.stringify(e));
                 }
-                this.log.info('Discovered %s %s - %s', device.deviceType, device.deviceModel, location.name, device.userDefinedDeviceName);
+                this.log.info(
+                  'Discovered %s %s - %s',
+                  device.deviceType,
+                  device.deviceModel,
+                  location.name,
+                  device.userDefinedDeviceName,
+                );
                 await this.createT9(device, locationId, this.firmware);
                 try {
                   await this.discoverRoomSensors(location.locationID, device);
                 } catch (e) {
-                  this.log.error('Failed to Find Room Sensors.', JSON.stringify(e.message));
+                  this.log.error('Failed to Find Room Sensor(s).', JSON.stringify(e.message));
                   this.log.debug(JSON.stringify(e));
                 }
               } else if (device.deviceModel.startsWith('T5')) {
-                this.log.info('Discovered %s %s - %s', device.deviceType, device.deviceModel, location.name, device.userDefinedDeviceName);
+                this.log.info(
+                  'Discovered %s %s - %s',
+                  device.deviceType,
+                  device.deviceModel,
+                  location.name,
+                  device.userDefinedDeviceName,
+                );
                 this.createT5(device, locationId);
               } else if (device.deviceModel.startsWith('D6')) {
-                this.log.info('Discovered %s %s - %s', device.deviceType, device.deviceModel, location.name, device.userDefinedDeviceName);
+                this.log.info(
+                  'Discovered %s %s - %s',
+                  device.deviceType,
+                  device.deviceModel,
+                  location.name,
+                  device.userDefinedDeviceName,
+                );
                 this.createT5(device, locationId);
               } else if (!device.DeviceModel) {
-                this.log.info('A LLC Device has been discovered with a deviceModel that does not start with T5, D6 or T9');
+                this.log.info(
+                  'A LLC Device has been discovered with a deviceModel that does not start with T5, D6 or T9',
+                );
               }
             } else if (device.deviceID.startsWith('TCC')) {
               this.log.info('A TCC Device has been discovered, Currently writing to Honeywell API does not work.');
               this.log.info(' Feel free to open an issue on GitHub https://git.io/JURI5');
               if (device.deviceModel.startsWith('Round')) {
-                this.log.info('Discovered %s %s - %s', device.deviceType, device.deviceModel, location.name, device.userDefinedDeviceName);
+                this.log.info(
+                  'Discovered %s %s - %s',
+                  device.deviceType,
+                  device.deviceModel,
+                  location.name,
+                  device.userDefinedDeviceName,
+                );
                 this.createRound(device, locationId);
               } else if (device.deviceModel.startsWith('Unknown')) {
-                this.log.info('Discovered %s %s - %s', device.deviceType, device.deviceModel, location.name, device.userDefinedDeviceName);
+                this.log.info(
+                  'Discovered %s %s - %s',
+                  device.deviceType,
+                  device.deviceModel,
+                  location.name,
+                  device.userDefinedDeviceName,
+                );
                 this.createTCC(device, locationId);
               } else if (!device.deviceModel) {
                 this.log.info(
@@ -440,8 +501,13 @@ export class HoneywellHomePlatform implements DynamicPlatformPlugin {
               if (sensorAccessory.accessoryAttribute) {
                 if (sensorAccessory.accessoryAttribute.type) {
                   if (sensorAccessory.accessoryAttribute.type.startsWith('IndoorAirSensor')) {
-                    this.log.info('Discovered Room Sensor groupId: %s, roomId: %s, accessoryId: %s', group.id, 
-                      sensorAccessory.roomId, sensorAccessory.accessoryId, sensorAccessory.accessoryAttribute.name);
+                    this.log.info(
+                      'Discovered Room Sensor groupId: %s, roomId: %s, accessoryId: %s',
+                      group.id,
+                      sensorAccessory.roomId,
+                      sensorAccessory.accessoryId,
+                      sensorAccessory.accessoryAttribute.name,
+                    );
                     this.createRoomSensors(device, locationId, sensorAccessory, group);
                     this.createRoomSensorThermostat(device, locationId, sensorAccessory, group);
                   }
@@ -454,19 +520,17 @@ export class HoneywellHomePlatform implements DynamicPlatformPlugin {
     }
   }
 
-
   private roomsensordisplaymethod() {
-    if (this.config.options ?.roompriority) {
+    if (this.config.options?.roompriority) {
       /**
        * Room Priority
        * This will display what room priority option that has been selected.
        */
       if (this.config.options.roompriority.thermostat) {
-        this.log.warn('Displaying Room Sensors as Thermostat(s).');
-        this.log.warn('You will have a Thermostat for Each Room Sensor so that you can set the priority of that Room.');
+        this.log.warn('Displaying Thermostat(s) for Each Room Sensor(s).');
       }
       if (!this.config.options.roompriority.thermostat) {
-        this.log.warn('Only Displaying Room Sensors.');
+        this.log.warn('Only Displaying Room Sensor(s).');
       }
     }
   }
@@ -484,8 +548,13 @@ export class HoneywellHomePlatform implements DynamicPlatformPlugin {
 
     if (existingAccessory) {
       // the accessory already exists
-      if (!this.config.options ?.thermostat ?.hide && device.isAlive) {
-        this.log.info('Restoring existing accessory from cache:', existingAccessory.displayName);
+      if (!this.config.options?.hide_device.includes(device.deviceID) && device.isAlive) {
+        this.log.info(
+          'Restoring existing accessory from cache:',
+          existingAccessory.displayName,
+          'DeviceID:',
+          device.deviceID,
+        );
 
         // if you need to update the accessory.context then you should run `api.updatePlatformAccessories`. eg.:
         existingAccessory.context.firmwareRevision = firmware;
@@ -494,14 +563,19 @@ export class HoneywellHomePlatform implements DynamicPlatformPlugin {
         // this is imported from `platformAccessory.ts`
         new T9thermostat(this, existingAccessory, locationId, device, firmware);
         this.log.debug(`T9 UDID: ${device.name}-${device.deviceID}-${device.deviceModel}`);
-      } else if (!device.isAlive || this.config.options ?.thermostat ?.hide) {
+      } else {
         this.unregisterPlatformAccessories(existingAccessory);
       }
-    } else if (!this.config.options ?.thermostat ?.hide) {
+    } else if (!this.config.options?.hide_device.includes(device.deviceID)) {
       // the accessory does not yet exist, so we need to create it
-      this.log.info('Adding new accessory:', `${device.name} ${device.deviceModel} ${device.deviceType}`);
-      this.log.debug(
-        `Registering new device: ${device.name} ${device.deviceModel} ${device.deviceType} - ${device.deviceID}`,
+      this.log.info(
+        'Adding new accessory:',
+        device.name,
+        'T9',
+        device.deviceModel,
+        device.deviceType,
+        'DeviceID:',
+        device.deviceID,
       );
 
       // create a new accessory
@@ -520,6 +594,19 @@ export class HoneywellHomePlatform implements DynamicPlatformPlugin {
       // link the accessory to your platform
       this.api.registerPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [accessory]);
       this.accessories.push(accessory);
+    } else {
+      this.log.error(
+        'Unable to Register new device:',
+        device.name,
+        'T9',
+        device.deviceModel,
+        device.deviceType,
+        'DeviceID:',
+        device.deviceID,
+      );
+      this.log.error(
+        'Check Config to see if DeviceID is being Hidden.',
+      );
     }
   }
 
@@ -532,8 +619,13 @@ export class HoneywellHomePlatform implements DynamicPlatformPlugin {
 
     if (existingAccessory) {
       // the accessory already exists
-      if (!this.config.options ?.thermostat ?.hide && device.isAlive) {
-        this.log.info('Restoring existing accessory from cache:', existingAccessory.displayName);
+      if (!this.config.options?.hide_device.includes(device.deviceID) && device.isAlive) {
+        this.log.info(
+          'Restoring existing accessory from cache:',
+          existingAccessory.displayName,
+          'DeviceID:',
+          device.deviceID,
+        );
 
         // if you need to update the accessory.context then you should run `api.updatePlatformAccessories`. eg.:
         //existingAccessory.context.firmwareRevision = findaccessories.accessoryAttribute.softwareRevision;
@@ -542,14 +634,19 @@ export class HoneywellHomePlatform implements DynamicPlatformPlugin {
         // this is imported from `platformAccessory.ts`
         new T5thermostat(this, existingAccessory, locationId, device);
         this.log.debug(`T5 UDID: ${device.name}-${device.deviceID}-${device.deviceModel}`);
-      } else if (!device.isAlive || this.config.options ?.thermostat ?.hide) {
+      } else {
         this.unregisterPlatformAccessories(existingAccessory);
       }
-    } else if (!this.config.options ?.thermostat ?.hide) {
+    } else if (!this.config.options?.hide_device.includes(device.deviceID)) {
       // the accessory does not yet exist, so we need to create it
-      this.log.info('Adding new accessory:', `${device.name} ${device.deviceModel} ${device.deviceType}`);
-      this.log.debug(
-        `Registering new device: ${device.name} ${device.deviceModel} ${device.deviceType} - ${device.deviceID}`,
+      this.log.info(
+        'Adding new accessory:',
+        device.name,
+        'T5',
+        device.deviceModel,
+        device.deviceType,
+        'DeviceID:',
+        device.deviceID,
       );
 
       // create a new accessory
@@ -567,6 +664,19 @@ export class HoneywellHomePlatform implements DynamicPlatformPlugin {
       // link the accessory to your platform
       this.api.registerPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [accessory]);
       this.accessories.push(accessory);
+    } else {
+      this.log.error(
+        'Unable to Register new device:',
+        device.name,
+        'T5',
+        device.deviceModel,
+        device.deviceType,
+        'DeviceID:',
+        device.deviceID,
+      );
+      this.log.error(
+        'Check Config to see if DeviceID is being Hidden.',
+      );
     }
   }
 
@@ -579,8 +689,13 @@ export class HoneywellHomePlatform implements DynamicPlatformPlugin {
 
     if (existingAccessory) {
       // the accessory already exists
-      if (!this.config.options ?.thermostat ?.hide && device.isAlive) {
-        this.log.info('Restoring existing accessory from cache:', existingAccessory.displayName);
+      if (!this.config.options?.hide_device.includes(device.deviceID) && device.isAlive) {
+        this.log.info(
+          'Restoring existing accessory from cache:',
+          existingAccessory.displayName,
+          'DeviceID:',
+          device.deviceID,
+        );
 
         // if you need to update the accessory.context then you should run `api.updatePlatformAccessories`. eg.:
         existingAccessory.context.firmwareRevision = device.thermostatVersion;
@@ -589,14 +704,19 @@ export class HoneywellHomePlatform implements DynamicPlatformPlugin {
         // this is imported from `platformAccessory.ts`
         new RoundThermostat(this, existingAccessory, locationId, device);
         this.log.debug(`Round UDID: ${device.name}-${device.deviceID}-${device.deviceModel}`);
-      } else if (!device.isAlive || this.config.options ?.thermostat ?.hide) {
+      } else {
         this.unregisterPlatformAccessories(existingAccessory);
       }
-    } else if (!this.config.options ?.thermostat ?.hide) {
+    } else if (!this.config.options?.hide_device.includes(device.deviceID)) {
       // the accessory does not yet exist, so we need to create it
-      this.log.info('Adding new accessory:', `${device.name} ${device.deviceModel} ${device.deviceType}`);
-      this.log.debug(
-        `Registering new device: ${device.name} ${device.deviceModel} ${device.deviceType} - ${device.deviceID}`,
+      this.log.info(
+        'Adding new accessory:',
+        device.name,
+        'Round',
+        device.deviceModel,
+        device.deviceType,
+        'DeviceID:',
+        device.deviceID,
       );
 
       // create a new accessory
@@ -614,6 +734,19 @@ export class HoneywellHomePlatform implements DynamicPlatformPlugin {
       // link the accessory to your platform
       this.api.registerPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [accessory]);
       this.accessories.push(accessory);
+    } else {
+      this.log.error(
+        'Unable to Register new device:',
+        device.name,
+        'Round',
+        device.deviceModel,
+        device.deviceType,
+        'DeviceID:',
+        device.deviceID,
+      );
+      this.log.error(
+        'Check Config to see if DeviceID is being Hidden.',
+      );
     }
   }
 
@@ -626,8 +759,13 @@ export class HoneywellHomePlatform implements DynamicPlatformPlugin {
 
     if (existingAccessory) {
       // the accessory already exists
-      if (!this.config.options ?.thermostat ?.hide && device.isAlive) {
-        this.log.info('Restoring existing accessory from cache:', existingAccessory.displayName);
+      if (!this.config.options?.hide_device.includes(device.deviceID) && device.isAlive) {
+        this.log.info(
+          'Restoring existing accessory from cache:',
+          existingAccessory.displayName,
+          'DeviceID:',
+          device.deviceID,
+        );
 
         // if you need to update the accessory.context then you should run `api.updatePlatformAccessories`. eg.:
         existingAccessory.context.firmwareRevision = device.thermostatVersion;
@@ -636,14 +774,19 @@ export class HoneywellHomePlatform implements DynamicPlatformPlugin {
         // this is imported from `platformAccessory.ts`
         new TCCthermostat(this, existingAccessory, locationId, device);
         this.log.debug(`TCC UDID: ${device.name}-${device.deviceID}-${device.deviceModel}`);
-      } else if (!device.isAlive || this.config.options ?.thermostat ?.hide) {
+      } else {
         this.unregisterPlatformAccessories(existingAccessory);
       }
-    } else if (!this.config.options ?.thermostat ?.hide) {
+    } else if (!this.config.options?.hide_device.includes(device.deviceID)) {
       // the accessory does not yet exist, so we need to create it
-      this.log.info('Adding new accessory:', `${device.name} TCC(${device.deviceModel}) ${device.deviceType}`);
-      this.log.debug(
-        `Registering new device: ${device.name} TCC(${device.deviceModel}) ${device.deviceType} - ${device.deviceID}`,
+      this.log.info(
+        'Adding new accessory:',
+        device.name,
+        'TCC',
+        device.deviceModel,
+        device.deviceType,
+        'DeviceID:',
+        device.deviceID,
       );
 
       // create a new accessory
@@ -661,6 +804,19 @@ export class HoneywellHomePlatform implements DynamicPlatformPlugin {
       // link the accessory to your platform
       this.api.registerPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [accessory]);
       this.accessories.push(accessory);
+    } else {
+      this.log.error(
+        'Unable to Register new device:',
+        device.name,
+        'TCC',
+        device.deviceModel,
+        device.deviceType,
+        'DeviceID:',
+        device.deviceID,
+      );
+      this.log.error(
+        'Check Config to see if DeviceID is being Hidden.',
+      );
     }
   }
 
@@ -673,8 +829,13 @@ export class HoneywellHomePlatform implements DynamicPlatformPlugin {
 
     if (existingAccessory) {
       // the accessory already exists
-      if (!this.config.options ?.leaksensor ?.hide && device.isAlive) {
-        this.log.info('Restoring existing accessory from cache:', existingAccessory.displayName);
+      if (!this.config.options?.hide_device.includes(device.deviceID) && device.isAlive) {
+        this.log.info(
+          'Restoring existing accessory from cache:',
+          existingAccessory.displayName,
+          'DeviceID:',
+          device.deviceID,
+        );
 
         // if you need to update the accessory.context then you should run `api.updatePlatformAccessories`. eg.:
         //existingAccessory.context.firmwareRevision = findaccessories.accessoryAttribute.softwareRevision;
@@ -684,14 +845,17 @@ export class HoneywellHomePlatform implements DynamicPlatformPlugin {
         // this is imported from `platformAccessory.ts`
         new LeakSensor(this, existingAccessory, locationId, device);
         this.log.debug(`Leak Sensor UDID: ${device.userDefinedDeviceName}-${device.deviceID}-${device.deviceClass}`);
-      } else if (!device.isAlive || this.config.options ?.leaksensor ?.hide) {
+      } else {
         this.unregisterPlatformAccessories(existingAccessory);
       }
-    } else if (!this.config.options ?.leaksensor ?.hide) {
+    } else if (!this.config.options?.hide_device.includes(device.deviceID)) {
       // the accessory does not yet exist, so we need to create it
-      this.log.info('Adding new accessory:', `${device.userDefinedDeviceName}  ${device.deviceClass}`);
-      this.log.debug(
-        `Registering new device: ${device.userDefinedDeviceName} ${device.deviceClass} - ${device.deviceID}`,
+      this.log.info(
+        'Adding new accessory:',
+        device.userDefinedDeviceName,
+        device.deviceClass,
+        'DeviceID:',
+        device.deviceID,
       );
 
       // create a new accessory
@@ -709,6 +873,17 @@ export class HoneywellHomePlatform implements DynamicPlatformPlugin {
       // link the accessory to your platform
       this.api.registerPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [accessory]);
       this.accessories.push(accessory);
+    } else {
+      this.log.error(
+        'Unable to Register new device:',
+        device.userDefinedDeviceName,
+        device.deviceClass,
+        'DeviceID:',
+        device.deviceID,
+      );
+      this.log.error(
+        'Check Config to see if DeviceID is being Hidden.',
+      );
     }
   }
 
@@ -718,7 +893,7 @@ export class HoneywellHomePlatform implements DynamicPlatformPlugin {
     sensorAccessory: sensorAccessory,
     group: T9groups,
   ) {
-    // Room Sensors
+    // Room Sensor(s)
     // this.log.info('createRoomSensors', device, locationId, sensorAccessory, group);
     const uuid = this.api.hap.uuid.generate(
       `${sensorAccessory.accessoryAttribute.name}-${sensorAccessory.accessoryAttribute.type}-${sensorAccessory.accessoryId}-RoomSensor`,
@@ -727,8 +902,13 @@ export class HoneywellHomePlatform implements DynamicPlatformPlugin {
 
     if (existingAccessory) {
       // the accessory already exists
-      if (device.isAlive && !this.config.options ?.roomsensor ?.hide) {
-        this.log.info('Restoring existing accessory from cache:', existingAccessory.displayName);
+      if (!this.config.options?.hide_device.includes(device.deviceID) && device.isAlive) {
+        this.log.info(
+          'Restoring existing accessory from cache:',
+          existingAccessory.displayName,
+          'DeviceID:',
+          device.deviceID,
+        );
 
         // if you need to update the accessory.context then you should run `api.updatePlatformAccessories`. eg.:
         //this.api.updatePlatformAccessories([existingAccessory]);
@@ -736,16 +916,18 @@ export class HoneywellHomePlatform implements DynamicPlatformPlugin {
         // create the accessory handler for the restored accessory
         // this is imported from `platformAccessory.ts`
         new RoomSensors(this, existingAccessory, locationId, device, sensorAccessory, group);
-
-      } else if (!device.isAlive || this.config.options ?.roomsensor ?.hide) {
+      } else {
         this.unregisterPlatformAccessories(existingAccessory);
       }
-    } else if (device.isAlive && !this.config.options ?.roomsensor ?.hide) {
+    } else if (!this.config.options?.hide_device.includes(device.deviceID) && device.isAlive) {
       // the accessory does not yet exist, so we need to create it
       this.log.info(
-        `Adding new accessory: ${sensorAccessory.accessoryAttribute.name} ${sensorAccessory.accessoryAttribute.type}`,
+        'Adding new accessory:',
+        sensorAccessory.accessoryAttribute.name,
+        sensorAccessory.accessoryAttribute.type,
+        'DeviceID:',
+        device.deviceID,
       );
-
 
       // create a new accessory
       const accessory = new this.api.platformAccessory(
@@ -764,6 +946,17 @@ export class HoneywellHomePlatform implements DynamicPlatformPlugin {
       // link the accessory to your platform
       this.api.registerPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [accessory]);
       this.accessories.push(accessory);
+    } else {
+      this.log.error(
+        'Unable to Register new device:',
+        sensorAccessory.accessoryAttribute.name,
+        sensorAccessory.accessoryAttribute.type,
+        'DeviceID:',
+        device.deviceID,
+      );
+      this.log.error(
+        'Check Config to see if DeviceID is being Hidden.',
+      );
     }
   }
 
@@ -800,21 +993,17 @@ export class HoneywellHomePlatform implements DynamicPlatformPlugin {
           // eslint-disable-next-line max-len
           `Room Sensor Thermostat UDID: ${sensorAccessory.accessoryAttribute.name}-${sensorAccessory.accessoryAttribute.type}-${sensorAccessory.accessoryId}-RoomSensorThermostat-${device.deviceID}`,
         );
-      } else if (!device.isAlive || !this.config.options ?.roompriority ?.thermostat) {
+      } else {
         this.unregisterPlatformAccessories(existingAccessory);
       }
     } else if (device.isAlive && this.config.options ?.roompriority ?.thermostat) {
       // the accessory does not yet exist, so we need to create it
       this.log.info(
         'Adding new accessory:',
-        `${sensorAccessory.accessoryAttribute.name} ${sensorAccessory.accessoryAttribute.type} Thermostat`,
-      );
-      this.log.debug(
-        'Registering new device: ',
         sensorAccessory.accessoryAttribute.name,
-        ' ',
         sensorAccessory.accessoryAttribute.type,
-        ' Thermostat - ',
+        'Thermostat.',
+        'DeviceID:',
         device.deviceID,
       );
 
@@ -839,13 +1028,22 @@ export class HoneywellHomePlatform implements DynamicPlatformPlugin {
       // link the accessory to your platform
       this.api.registerPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [accessory]);
       this.accessories.push(accessory);
+    } else {
+      this.log.warn(
+        'Room Priority is not set, New device will not be Registered: ',
+        sensorAccessory.accessoryAttribute.name,
+        sensorAccessory.accessoryAttribute.type,
+        'Thermostat',
+        'DeviceID:',
+        device.deviceID,
+      );
     }
   }
 
   public unregisterPlatformAccessories(existingAccessory: PlatformAccessory) {
     // remove platform accessories when no longer present
     this.api.unregisterPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [existingAccessory]);
-    this.log.info('Removing existing accessory from cache:', existingAccessory.displayName);
+    this.log.warn('Removing existing accessory from cache:', existingAccessory.displayName);
   }
 
   public locationinfo(location: location) {