+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file. This project uses [Semantic Versioning](https://semver.org/)
 
+## [Version 8.3.0](https://github.com/donavanbecker/homebridge-honeywell-home/compare/v8.2.0...v8.3.0) (2021-11-04)
+
+### Changes
+
+- Allow for Hiding Devices based off of `DeviceID` instead of `DeviceType`.
+- Fix issue with Leak Sensors not updating new sensor data.
+- Changing `ttl` code to `refreshRate`, Configs will need to be update to not use default Refresh Rate.
+- Code Clean up.
+
 ## [Version 8.2.0](https://github.com/donavanbecker/homebridge-honeywell-home/compare/v8.1.1...v8.2.0) (2020-12-25)
 
 ### Changes